#59 Move exception wrapper stuff to exception-handler
