+++ services/employee/src/main/java/com/senla/worklog/reminder/employee/application/exception/wrapper/EmployeeNotFoundExceptionWrapper.java
@@ -1,6 +1,7 @@
 package com.senla.worklog.reminder.employee.application.exception.wrapper;
 
-import com.senla.worklog.reminder.employee.application.exception.ApplicationException;
+import com.senla.common.exception.ApplicationException;
+import com.senla.common.exception.wrapper.ExceptionWrapper;
 import com.senla.worklog.reminder.employee.application.exception.ResourceNotFoundException;
 import com.senla.worklog.reminder.employee.domain.exception.EmployeeNotFoundException;
 import org.springframework.stereotype.Component;
