+++ services/employee/src/test/java/com/senla/worklog/reminder/employee/adapter/in/rest/exception/handler/RestAdapterExceptionHandlerTest.java
@@ -1,8 +1,8 @@
 package com.senla.worklog.reminder.employee.adapter.in.rest.exception.handler;
 
+import com.senla.common.exception.UnexpectedApplicationException;
 import com.senla.common.exception.handler.model.AttributeApiSubError;
 import com.senla.worklog.reminder.employee.application.exception.ResourceNotFoundException;
-import com.senla.worklog.reminder.employee.application.exception.UnexpectedApplicationException;
 import com.senla.worklog.reminder.employee.application.exception.UniqueConstraintViolationException;
 import com.senla.worklog.reminder.employee.domain.exception.DuplicateAttributeException;
 import com.senla.worklog.reminder.employee.domain.exception.EmployeeNotFoundException;
