+++ services/exception-handler/build.gradle
@@ -22,6 +22,7 @@ publishing {
 
             dependencies {
                 implementation 'org.springframework.boot:spring-boot-starter-web:2.7.8'
+                implementation 'org.springframework.boot:spring-boot-starter-aop:2.7.8'
             }
         }
     }
@@ -38,6 +39,7 @@ dependencyManagement {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-aop'
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
