+++ services/exception-handler/src/main/java/com/senla/common/exception/ApplicationException.java
@@ -1,4 +1,4 @@
-package com.senla.worklog.reminder.employee.application.exception;
+package com.senla.common.exception;
 
 /**
  * The ApplicationException class is an abstract base class for custom application exceptions. It extends
